有锁住的动作 但是在指定点抖动很大 后面改一下方式 增加速度环 速度环反馈给位置环

+++ FLASH/flash_helper.h
@@ -31,11 +31,11 @@
 #define ADDR_FLASH_APP_PAGE						(ADDR_FLASH_START_PAGE+(APP_BASE*FLASH_PAGE_SIZE))
 #define ADDR_FLASH_EEPROM_PAGE					(ADDR_FLASH_START_PAGE+(EEPROM_BASE*FLASH_PAGE_SIZE)) 
 // Functions
-FLASH_Status write_data(uint32_t base, uint8_t *data, uint32_t len);
+FLASH_Status write_data(uint32_t base, uint32_t *data, uint32_t len);
 FLASH_Status erase_page(uint32_t start_addr_page,uint32_t new_app_size);
 FLASH_Status flash_helper_erase_new_app(uint32_t new_app_size);
 FLASH_Status flash_helper_erase_bootloader(void);
-FLASH_Status flash_helper_write_new_app_data(uint32_t offset, uint8_t *data, uint32_t len);
+FLASH_Status flash_helper_write_new_app_data(uint32_t offset, uint32_t *data, uint32_t len);
 void flash_helper_jump_to_bootloader(void);
 
 #endif /* FLASH_HELPER_H_ */
